migrating errors corrected 24/5

+++ database/migrations/2022_04_27_084759_create_holidays_table.php
@@ -15,10 +15,10 @@ return new class extends Migration
     {
         Schema::create('holidays', function (Blueprint $table) {
            // $table->id();
-          /*  $table->UnSignedInteger('Holiday_ID');*/
+          /*  $table->unsignedInteger('holiday_id');*/
             $table->date('holiday_fall_date');
             $table->string('holiday_description');
-            $table->unsignedInteger('department_id');
+            $table->unsignedBigInteger('department_id');
             $table->timestamps();
         });
